Added styling to addcitizen.tsx and [grade].tsx need to finish styling for removecitizen.tsx and then rework logic to be able to select multiple at a time
