+++ app/auth/profile/organisation/grade/[grade].tsx
@@ -1,0 +1,134 @@
+import { Ionicons } from "@expo/vector-icons";
+import { router, useLocalSearchParams } from "expo-router";
+import { Fragment, useState } from "react";
+import { SafeAreaView, StyleSheet, Text, View } from "react-native";
+import { FlatList } from "react-native-gesture-handler";
+import IconButton from "../../../../../components/IconButton";
+import useClasses, { ClassDTO } from "../../../../../hooks/useClasses";
+import { colors, ScaleSize, ScaleSizeH, ScaleSizeW, SharedStyles } from "../../../../../utils/SharedStyles";
+import SearchBar from "../../../../../components/SearchBar";
+
+const ViewGrade = () => {
+  const { grade } = useLocalSearchParams();
+  const parsedID = Number(grade);
+  const [searchedCitizens, setSearchedCitizens] = useState<string>("");
+  const { data, error, isLoading } = useClasses(parsedID);
+  const currentGrade = data?.grades.find((grade) => grade.id === parsedID);
+
+  if (isLoading)
+    return (
+      <View>
+        <Text>Loading...</Text>
+      </View>
+    );
+  if (error)
+    return (
+      <View>
+        <Text>Error loading class data</Text>
+      </View>
+    );
+
+  //Sorts citizens alphabetically
+  const sortedCitizen = (data: ClassDTO) => {
+    return data.citizens
+      .filter((citizen) =>
+        `${citizen.firstName} ${citizen.lastName}`.toLowerCase().startsWith(searchedCitizens.toLowerCase())
+      )
+      .sort((a, b) => a.firstName.localeCompare(b.firstName) || a.lastName.localeCompare(b.lastName));
+  };
+
+  return (
+    <Fragment>
+      <FlatList
+        data={currentGrade ? sortedCitizen(currentGrade) : []}
+        keyExtractor={(item) => item.id.toString()}
+        contentContainerStyle={styles.citizenView}
+        numColumns={2}
+        renderItem={({ item }) => (
+          <Text style={styles.citizenName}>{item.firstName + " " + item.lastName}</Text>
+        )}
+        ListEmptyComponent={<Text>Ingen elever tilføjet til denne klasse</Text>}
+        ListHeaderComponent={
+          <View style={styles.searchbarStyle}>
+            <View style={{ alignItems: "center" }}>
+              <Text style={styles.gradeName}>{currentGrade?.name ?? "Class"}</Text>
+              <View style={styles.ActionView}>
+                <IconButton onPress={() => router.back()} absolute={false}>
+                  <Ionicons name={"exit-outline"} size={ScaleSize(30)} />
+                </IconButton>
+                {/* Edit Grade */}
+                <IconButton onPress={() => {}} absolute={false}>
+                  <Ionicons name={"create-outline"} size={ScaleSize(30)} />
+                </IconButton>
+                {/* Collective Weekplan */}
+                <IconButton onPress={() => {}} absolute={false}>
+                  <Ionicons name={"calendar-outline"} size={ScaleSize(30)} />
+                </IconButton>
+              </View>
+              <View style={styles.ActionView}>
+                <IconButton
+                  onPress={() => {
+                    router.push({
+                      pathname: "/auth/profile/organisation/grade/addcitizen",
+                      params: { gradeId: grade },
+                    });
+                  }}
+                  absolute={false}>
+                  <Ionicons name={"person-add-outline"} size={ScaleSize(30)} />
+                </IconButton>
+                <IconButton
+                  onPress={() => {
+                    router.push({
+                      pathname: "/auth/profile/organisation/grade/removecitizen",
+                      params: { gradeId: grade },
+                    });
+                  }}
+                  absolute={false}>
+                  <Ionicons name={"person-remove-outline"} size={ScaleSize(30)} />
+                </IconButton>
+              </View>
+            </View>
+            <SearchBar value={searchedCitizens} onChangeText={setSearchedCitizens} />
+          </View>
+        }
+      />
+    </Fragment>
+  );
+};
+
+const styles = StyleSheet.create({
+  gradeName: {
+    fontSize: ScaleSize(40),
+    fontWeight: "bold",
+    marginTop: ScaleSizeH(25),
+    marginBottom: ScaleSize(10),
+  },
+  ActionView: {
+    ...SharedStyles.flexRow,
+    gap: ScaleSizeW(10),
+  },
+  citizenView: {
+    width: "100%",
+    alignItems: "center",
+  },
+  searchbarStyle: {
+    width: "90%",
+    minWidth: "90%",
+  },
+  citizenName: {
+    fontSize: ScaleSize(20),
+    textAlign: "center",
+    paddingVertical: ScaleSizeH(25),
+    borderRadius: 15,
+    backgroundColor: colors.lightBlue,
+    width: ScaleSizeW(300),
+    marginVertical: ScaleSizeH(5),
+    marginHorizontal: ScaleSize(5),
+  },
+  sheetContent: {
+    alignItems: "center",
+    paddingVertical: ScaleSize(20),
+  },
+});
+
+export default ViewGrade;
